+++ test/run_all_versions.sh
@@ -116,6 +116,7 @@ SUITES=(
 	import_export_privilege
 	index_delete_liveness
 	index_fetch_penalty_width
+	index_fetch_penalty_crossover
 	index_only
 	inheritance
 	int8_agg_int128